[Fix] tl.dot lhs & rhs dtype

+++ transformer_engine/pytorch/triton/sage_attn_v1.py
@@ -333,14 +333,16 @@ def _attn_bwd_dkdv(dk, dv,
             pT = tl.where(mask, pT, 0.0)
         do = tl.load(do_ptrs)
         # Compute dV.
-        ppT = pT.to(tl.float16)
+        # ppT = pT.to(tl.float16)
+        ppT = pT.to(do.dtype) 
         dv += tl.dot(ppT, do)
         # D (= delta) is pre-divided by ds_scale.
         Di = tl.load(D + offs_m)
         # Compute dP and dS.
         dpT = tl.dot(v, tl.trans(do)).to(tl.float32)
         dsT = pT * (dpT - Di[None, :])
-        dsT = dsT.to(tl.float16)
+        # dsT = dsT.to(tl.float16)
+        dsT = dsT.to(qT.dtype) 
         dk += tl.dot(dsT, tl.trans(qT))
         # Increment pointers.
         curr_m += step_m
@@ -386,7 +388,8 @@ def _attn_bwd_dq(dq, q, K, V,
         # Compute dP and dS.
         dp = tl.dot(do, vT).to(tl.float32)
         ds = p * (dp - Di[:, None])
-        ds = ds.to(tl.float16)
+        # ds = ds.to(tl.float16)
+        ds = ds.to(kT.dtype) 
         # Compute dQ.
         # NOTE: We need to de-scale dq in the end, because kT was pre-scaled.
         dq += tl.dot(ds, tl.trans(kT))
